Add module preprocessor and fix linter issues
