+++ .github/linters/sun_checks.xml
@@ -120,9 +120,10 @@
     <module name="IllegalImport"/>
     <!-- defaults to sun.* packages -->
     <module name="RedundantImport"/>
-    <module name="UnusedImports">
+    <!-- Commenting for now, not suitable for generated events -->
+    <!--<module name="UnusedImports">
       <property name="processJavadoc" value="false"/>
-    </module>
+    </module>-->
 
     <!-- Checks for Size Violations.                    -->
     <!-- See https://checkstyle.org/config_sizes.html -->
