+++ pom.xml
@@ -86,6 +86,7 @@
 	</scm>
 
 	<modules>
+		<module>preprocessor</module>
 		<module>generator</module>
 		<module>sdk</module>
 	</modules>
