+++ sdk/src/main/java/dev/cdevents/spec/schemas/SchemaLoader.java
@@ -1,15 +1,11 @@
 package dev.cdevents.spec.schemas;
 
-import com.networknt.schema.JsonSchema;
-import com.networknt.schema.JsonSchemaFactory;
-import com.networknt.schema.SpecVersion;
-import dev.cdevents.CDEvents;
 import dev.cdevents.constants.CDEventConstants;
 import dev.cdevents.exception.CDEventsException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.io.*;
+import java.io.BufferedReader;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.stream.Collectors;
@@ -19,8 +15,6 @@ import java.util.stream.Stream;
  * The purpose of this class is to serve as a resource anchor for loading event schemas.
  */
 public final class SchemaLoader {
-    private static Logger log = LoggerFactory.getLogger(SchemaLoader.class);
-
     private SchemaLoader() {
 
     }
@@ -34,44 +28,17 @@ public final class SchemaLoader {
      * @return schema contents
      */
     public static String loadSchema(String schema) {
-        StringBuilder stringBuilder = new StringBuilder();
-        String eventSchema = loadSchema(schema, "dev/cdevents/spec/schemas/", CDEventConstants.SCHEMA_FOLDER);
-        stringBuilder.append(eventSchema);
-        stringBuilder.append(System.lineSeparator());
-
-        /*String embeddedLinksArray = loadSchema("embeddedlinksarray.json", "dev/cdevents/spec/schemas/links/", CDEventConstants.SCHEMA_LINKS_FOLDER);
-        stringBuilder.append(embeddedLinksArray);
-        stringBuilder.append(System.lineSeparator());
-
-        String embeddedLinkEnd = loadSchema("embeddedlinkend.json", "dev/cdevents/spec/schemas/links/", CDEventConstants.SCHEMA_LINKS_FOLDER);
-        stringBuilder.append(embeddedLinkEnd);
-        stringBuilder.append(System.lineSeparator());
-
-        String embeddedLinkPath = loadSchema("embeddedlinkpath.json", "dev/cdevents/spec/schemas/links/", CDEventConstants.SCHEMA_LINKS_FOLDER);
-        stringBuilder.append(embeddedLinkPath);
-        stringBuilder.append(System.lineSeparator());
-
-        String embeddedLinkRelation = loadSchema("embeddedlinkrelation.json", "dev/cdevents/spec/schemas/links/", CDEventConstants.SCHEMA_LINKS_FOLDER);
-        stringBuilder.append(embeddedLinkRelation);
-        stringBuilder.append(System.lineSeparator());*/
-
-        return stringBuilder.toString();
-    }
-
-    private static String loadSchema(String schema, String classPath, String fileSystemPath){
         try {
             // load schema from classpath first
             InputStream inputStream = SchemaLoader.class.getClassLoader()
-                    .getResourceAsStream(classPath + schema);
+                .getResourceAsStream("dev/cdevents/spec/schemas/" + schema);
             if (null != inputStream) {
-                log.info("loadSchema from classPath -> " +classPath);
                 return new BufferedReader(new InputStreamReader(inputStream))
-                        .lines().collect(Collectors.joining("\n"));
+                    .lines().collect(Collectors.joining("\n"));
             }
 
             // load from filesystem next
-            try (Stream<String> lines = Files.lines(Paths.get(fileSystemPath + "/" + schema))) {
-                log.info("loadSchema from fileSystemPath -> " +fileSystemPath);
+            try (Stream<String> lines = Files.lines(Paths.get(CDEventConstants.SCHEMA_FOLDER + "/" + schema))) {
                 return lines.collect(Collectors.joining("\n"));
             }
         } catch (Exception e) {
@@ -79,4 +46,3 @@ public final class SchemaLoader {
         }
     }
 }
-
